test(dev): align e2e suite with Vite and parallelize playgrounds (#9759)
## Summary

Aligns the dev-server browser e2e suite (`packages/test-dev-server/tests`) with Vite's playground harness and **optimizes test performance by running playgrounds in parallel**.

## `vitest.config.e2e.mts` — match Vite's `vitest.config.e2e.ts`

- **Enable file parallelism.** Vite sets neither `pool` nor `fileParallelism`, so it runs playgrounds in parallel on vitest's defaults. We were forcing `fileParallelism: false`; dropping it (and the explicit `pool`) opts into the same default `forks` pool + parallel execution. The `forks` pool matters for us: each spec runs an in-process rust/napi dev engine that needs process isolation.
- **Drop `retry`** (vitest defaults to `0`, like Vite). The product flake that once justified a retry crutch is fixed (`compareContentsForPolling`, #9736).
- Add `reporters: 'dot'` and the `PWDEBUG → Infinity` timeout. Keep the rolldown-only `env` (`RUST_BACKTRACE`/`RD_LOG`).

## `hmr-full-bundle-mode`: merge 4 specs → 1 file

Parallelism runs **spec files** concurrently, each forking its own dev server against the playground's shared `playground-temp/<name>/` copy. Multiple spec files in one playground are safe **only when their scenarios are disjoint** — own page/DOM/source files, no spec editing a file another spec's page imports. That is how Vite runs its multi-spec playgrounds (and how `lazy-compilation` already works).

`hmr-full-bundle-mode` can't be disjoint: it's *full-bundle* mode with one shared entry (`main.js` imports every scenario's module, `index.html` renders them all), so every spec's server bundles every file and an edit by one spec would rebuild a sibling's page mid-test. Its four `__tests__/*.spec.ts` are therefore merged into one `hmr-full-bundle-mode.spec.ts` — now `describe` blocks that run sequentially, each restoring what it edits.

`lazy-compilation` is **left unchanged**: its four specs already use disjoint scenarios (separate DOM nodes and source folders, lazily compiled so they never warm each other), so they run safely in parallel as separate spec files.

## Verification

Local: `pnpm run test:browser` — 5 test files / 14 tests pass in parallel with no retry.
